Complete rewrite of char_seq.py
